OpenMX: parse stress tensor
This is quite tricky, because for unknown reasons the stress tensor is
not printed into the standard mainfile *.out but only into the standard
output. This means, we have to try all files in the directory and check
first few lines to see if anything resembles the OpenMX standard output,
than do a few consistency check to be extra sure this is a correct
stdout corresponding to the mainfile. The parsing itself is trivial.

v2: minor fixes by Pavel Ondračka